+++ common-db/src/test/java/cn/featherfly/common/db/dialect/DialectTest.java
@@ -225,4 +225,6 @@ public abstract class DialectTest {
 
     abstract void testUpsertBatch();
 
+    abstract void testDeleteFrom();
+
 }