+++ common-db/src/test/java/cn/featherfly/common/db/dialect/PostgresqlDialectTest.java
@@ -208,4 +208,19 @@ public class PostgresqlDialectTest extends DialectTest {
         System.out.println(sql);
         assertEquals(sql, "INSERT INTO \"user\" (\"id\", \"name\", \"age\") VALUES (?, ?, ?),(?, ?, ?)");
     }
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    @Test
+    void testDeleteFrom() {
+        String sql = dialect.buildDeleteFromSql("user");
+        System.out.println(sql);
+        assertEquals(sql, "DELETE FROM \"user\"");
+
+        sql = dialect.buildDeleteFromSql("user", "u");
+        System.out.println(sql);
+        assertEquals(sql, "DELETE FROM \"user\" \"u\"");
+    }
 }