+++ common-db/change-log.md
@@ -5,6 +5,7 @@
 4. 优化SqlTypeMappingManager关于JavaSqlTypeMapper的逻辑
 5. 加入JdbcPropertyMapping
 6. 加入FieldValueOperator，用于优化Entity的数据设置与解析不走SqlTypeMappingManager
+7. Dialect加入buildDeleteFromSql方法,SqlDeleteFromBasicBuilder加入tableAlias支持
 
 # 0.5.5 2022-8-11
 1. Dialect加入getKeywordLike(QueryPolicy)、getKeywordEq(QueryPolicy)、keywordsCase()方法